+++ proyecto/Plot.hpp
@@ -7,7 +7,7 @@ class Plot {
 
 public:
 
-    static void plot(vector<vector<int> > world, vector<pair<int, int> > path, pair<int, int> start_pos, pair<int, int> end_pos);
+    static void plot(vector<vector<double> > world, vector<pair<int, int> > path, pair<int, int> start_pos, pair<int, int> end_pos);
 
 };
